Big internal changes
- Return C struct instead of objc object
- Use Enums for error status,
- Use bindgen to generate bindings (I trust it's output more
  than I trust my own skills, also I'll learn something looking
  at the generated bindings)
- Removes objc as a dependency. At the interface between objc and
  rust, we only deal with C structs.
- Truncate floats inside of C, and not inside of rust code. Idk

+++ src/error.rs
@@ -4,7 +4,7 @@ use std::{error, fmt};
 pub enum Error {
     NotEnabled,
     NotReturned,
-    Stale(i64),
+    Stale(i32),
     Unknown,
 }
